+++ src/components/budgets/ModalPDF/index.jsx
@@ -1,7 +1,7 @@
 import { useDolarExangeRate } from "@/api/external";
 import { IconedButton } from "@/common/components/buttons";
 import { ButtonsContainer, Flex, FlexColumn, Input } from "@/common/components/custom";
-import { COLORS, ICONS } from "@/common/constants";
+import { BUTTON_TEXTS, COLORS, ICONS } from "@/common/constants";
 import { OnlyPrint } from "@/components/layout";
 import { useEffect, useRef, useState } from "react";
 import { useReactToPrint } from "react-to-print";
@@ -157,7 +157,7 @@ const ModalPDF = ({
           <Modal.Actions>
             <ButtonsContainer width="100%">
               <IconedButton
-                text="Cancelar"
+                text={BUTTON_TEXTS.CANCEL}
                 icon={ICONS.CANCEL}
                 color={COLORS.RED}
                 onClick={() => onClose(false)}