+++ src/components/users/UserForm/index.jsx
@@ -1,7 +1,7 @@
 import { SubmitAndRestore } from "@/common/components/buttons";
 import { FieldsContainer, Form, FormField } from "@/common/components/custom";
 import { DropdownControlled, NumberControlled, TextAreaControlled, TextControlled } from "@/common/components/form";
-import { RULES, SHORTKEYS } from "@/common/constants";
+import { ERROR_MESSAGES, FIELD_LABELS, RULES, SHORTKEYS } from "@/common/constants";
 import { validateEmail } from "@/common/utils";
 import { getPastDate } from "@/common/utils/dates";
 import { useKeyboardShortcuts } from "@/hooks";
@@ -69,7 +69,7 @@ const UserForm = forwardRef(({
               label="Usuario"
               placeholder="martinb@hotmail.com"
               rules={{
-                required: "Este campo es obligatorio.",
+                required: ERROR_MESSAGES.REQUIRED_FIELD_ALT,
                 validate: {
                   email: (value) => validateEmail(value) || "El correo electrónico no es válido.",
                 },
@@ -100,7 +100,7 @@ const UserForm = forwardRef(({
           <FormField flex="1">
             <TextControlled
               name="firstName"
-              label="Nombre"
+              label={FIELD_LABELS.NAME}
               placeholder="Martín"
               rules={RULES.REQUIRED}
               disabled={!isUpdating && view}
@@ -134,7 +134,7 @@ const UserForm = forwardRef(({
               scrollableYearDropdown
               yearDropdownItemNumber={80}
               rules={{
-                required: "Campo requerido.",
+                required: ERROR_MESSAGES.REQUIRED_FIELD,
                 validate: (value) => {
                   const today = new Date();
                   const minBirthDate = new Date();
@@ -200,7 +200,7 @@ const UserForm = forwardRef(({
           <FormField flex="1" />
         </FieldsContainer>
         <FieldsContainer>
-          <TextAreaControlled name="comments" label="Comentarios" placeholder="Martin Bueno no era un cliente?" readOnly={!isUpdating && view} />
+          <TextAreaControlled name="comments" label={FIELD_LABELS.COMMENTS} placeholder="Martin Bueno no era un cliente?" readOnly={!isUpdating && view} />
         </FieldsContainer>
         {(isUpdating || !view) && (
           <SubmitAndRestore